@@ -22,9 +22,9 @@
     "babel": "^6.23.0",
     "babel-loader": "^8.0.6",
     "file-loader": "^4.2.0",
+    "pegjs-loader": "^0.5.5",
     "webpack-cli": "^3.3.9",
-    "webpack-dev-server": "^4.11.1",
-    "pegjs-loader": "^0.5.5"
+    "webpack-dev-server": "^4.11.1"
   },
   "dependencies": {
     "@babel/plugin-proposal-class-properties": "^7.5.5",
@@ -32,8 +32,8 @@
     "express": "^4.17.1",
     "pegjs": "^0.10.0",
     "pug": "^3.0.2",
-    "react": "^16.10.2",
-    "react-dom": "^16.10.2",
-    "webpack": "^4.41.2"
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "webpack": "^5.99.5"
   }
 }
