@@ -13,6 +13,7 @@ export class Display extends React.Component {
     constructor(props) {
         super(props);
 
+        this.display = React.createRef();
         this.chip = props.chip;
     }
 
@@ -37,7 +38,7 @@ export class Display extends React.Component {
     }
 
     componentDidMount() {
-        this.ctx = this.refs.display.getContext('2d');
+        this.ctx = this.display.current.getContext('2d');
 
         this.grad = this.ctx.createLinearGradient(0, 0, 0, displayHeight);
         this.grad.addColorStop(0.0, `rgba(${background.r},${background.g},${background.b},1)`);
@@ -48,7 +49,7 @@ export class Display extends React.Component {
     }
 
     render() {
-        return <canvas id="emu-display" ref="display" width={displayWidth} height={displayHeight}></canvas>;
+        return <canvas id="emu-display" ref={this.display} width={displayWidth} height={displayHeight}></canvas>;
     }
 }
 
